missing Q_OBJECTs
